finish

+++ src/logics/projects.logics.ts
@@ -5,7 +5,7 @@ import { client } from "../database";
 import { internalServerError, keysMissing, outFormat } from "../errors/common.errors";
 import { iValidade } from "../interfaces/common.interfaces";
 import { iProjectCreate, iProjectRequest, iProjectResult, iProjectTechnologiesResult, iProjectWithTechResult, iTechResult, iValidateCreateTech } from "../interfaces/projects.interfaces";
-import { validateCreateProject, validateCreateTech } from "../validates/projects.validates";
+import { validateCreateProject, validateCreateTech, validateUpdateProject } from "../validates/projects.validates";
 
 export async function showProjects(req: Request, res: Response): Promise<Response> {
     const queryString: string = `--sql
@@ -21,22 +21,21 @@ export async function showProjects(req: Request, res: Response): Promise<Respons
             ti."id" "technologyId",
             ti."name" "technologyName"
         FROM
-            projects_technologies pt
-        JOIN 
-            projects pj ON pt."projectId" = pj.id
-        JOIN 
+            projects pj
+        LEFT JOIN 
+            projects_technologies pt ON pt."projectId" = pj.id
+        LEFT JOIN 
             technologies ti ON pt."technologyId" = ti.id;
     `;
 
     const queryResult: iProjectWithTechResult = await client.query(queryString);
-    console.log(queryResult)
     
     return res.status(200).json(queryResult.rows);
 }
 
 export async function findProject(req: Request, res: Response): Promise<Response> {
     const idProject: number = req.idProject!;
-    console.log(idProject)
+    
     const queryString: string = `--sql
         SELECT
             pj."id" "projectId",
@@ -50,13 +49,13 @@ export async function findProject(req: Request, res: Response): Promise<Response
             ti."id" "technologyId",
             ti."name" "technologyName"
         FROM
-            projects_technologies pt
-        JOIN 
-            projects pj ON pt."projectId" = pj.id
-        JOIN 
+            projects pj
+        LEFT JOIN 
+            projects_technologies pt ON pt."projectId" = pj.id
+        LEFT JOIN 
             technologies ti ON pt."technologyId" = ti.id
         WHERE
-            pt.id = $1;
+            pj.id = $1;
     `;
 
     const queryConfig: QueryConfig = {
@@ -93,14 +92,36 @@ export async function deleteTech(req: Request, res: Response): Promise<Response>
     const idProject: number = req.idProject!;
     const idTech: number = req.idTech!;
  
-    const queryString: string = `--sql
-        DELETE FROM
+    let queryString: string = `--sql
+        SELECT 
+            *
+        FROM
             projects_technologies pj
         WHERE
             pj."projectId" = $1 and pj."technologyId" = $2;
     `; 
 
-    const queryConfig: QueryConfig = {
+    let queryConfig: QueryConfig = {
+        text: queryString,
+        values: [idProject, idTech]
+    }
+
+    const queryResult = await client.query(queryConfig);
+
+    if (queryResult.rowCount === 0) {
+        return res.status(404).json({
+            message: `Not found tech`
+        });
+    }
+
+    queryString = `--sql
+        DELETE FROM
+            projects_technologies pj
+        WHERE
+            pj."projectId" = $1 and pj."technologyId" = $2;
+    `;
+
+    queryConfig = {
         text: queryString,
         values: [idProject, idTech]
     }
@@ -111,8 +132,6 @@ export async function deleteTech(req: Request, res: Response): Promise<Response>
 }
 
 export async function createProject(req: Request, res: Response): Promise<Response> {
-    let idProject: number = 0;
-    let dataResponse: iProjectCreate | {} = {};
     const resultValidate: iValidade = validateCreateProject(req);
     if (!resultValidate.status) {
         if (resultValidate.keysMissing.length > 0) {
@@ -122,7 +141,6 @@ export async function createProject(req: Request, res: Response): Promise<Respon
     }
 
     const projectData: iProjectRequest = req.projectData!;
-    const idDev: number = req.idDev!;
 
     let queryString: string = format(`--sql
         INSERT INTO
@@ -136,7 +154,6 @@ export async function createProject(req: Request, res: Response): Promise<Respon
     );
     try {
         const queryResult: iProjectResult = await client.query(queryString);
-        idProject = queryResult.rows[0].id;
         return res.status(201).json(queryResult.rows[0]);
     } catch(error: any) {
         return internalServerError(res, error.message);
@@ -216,4 +233,36 @@ export async function createTech(req: Request, res: Response): Promise<Response>
     const queryResponse: iProjectWithTechResult = await client.query(queryConfig);
 
     return res.status(201).json(queryResponse.rows[0]);
+}
+
+export async function updateProject(req: Request, res: Response): Promise<Response> {
+    const resultValidate: iValidade = validateUpdateProject(req);
+    if (!resultValidate.status) {
+        if (resultValidate.keysMissing.length > 0) {
+            return keysMissing(res, resultValidate.keysMissing);
+        }
+        return outFormat(res);
+    }
+    const idProject: number = req.idProject!;
+    const projectData: iProjectRequest = req.projectData!
+
+    const queryString: string = format(`--sql
+        UPDATE 
+            projects
+        SET(%I) = ROW(%L)
+        WHERE 
+            id = %s
+        RETURNING *;
+    `,
+    Object.keys(projectData),
+    Object.values(projectData),
+    idProject
+    );
+
+    try {
+        const queryResult: iProjectResult = await client.query(queryString);
+        return res.status(200).json(queryResult.rows);
+    } catch(error: any) {
+        return internalServerError(res, error.message);
+    }
 }
